Add pass to check for CAFs with basic values
For example:

```clean
x =: 5
```

It is faster to just push the value to the stack again (i.e. `x = 5` or
`x :== 5`) than to use a CAF. With a CAF, each usage of `x` would check
whether it has been evaluated before, and store a reference in the data
segment after it has been evaluated for the first time.