Add failing test for ignoring didClose notifications
The correct behaviour is that the language server does *not* respond to
textDocument/didClose notifications. This is difficult to test.

With the setup in this commit, we allow 500ms for the language server to
still send notifications. After that, we assume that no notifications
will be sent.

Unfortunately, it is not possible to terminate a process and then read
any remaining data in the pipes, because the pipes have been cleaned up.
This means we need to do a non-blocking read just before terminating the
process.